ISSUE #4780 - update container revision names to include container
