+++ frontend/src/v5/ui/routes/dashboard/projects/containers/uploadContainerRevisionForm/uploadList/uploadList.component.tsx
@@ -15,7 +15,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { UploadItemFields } from '@/v5/store/containers/containers.types';
+import { UploadItemFields } from '@/v5/store/containerRevisions/containerRevisions.types';
 import { useContext } from 'react';
 import { SortingDirection } from '@components/dashboard/dashboardList/dashboardList.types';
 import { useOrderedList } from '@components/dashboard/dashboardList/useOrderedList';
