+++ frontend/src/v5/ui/routes/dashboard/projects/containers/uploadContainerRevisionForm/uploadList/uploadListItem/uploadListItem.component.tsx
@@ -25,7 +25,8 @@ import { UploadListItemRevisionTag } from './components/uploadListItemRevisionTa
 import { UploadListItemButton } from './uploadListItem.styles';
 import { useFormContext } from 'react-hook-form';
 import { useEffect } from 'react';
-import { IContainer, UploadItemFields, UploadStatuses } from '@/v5/store/containers/containers.types';
+import { IContainer, UploadStatuses } from '@/v5/store/containers/containers.types';
+import { UploadItemFields } from '@/v5/store/containerRevisions/containerRevisions.types';
 import { ContainersActionsDispatchers } from '@/v5/services/actionsDispatchers';
 import { UploadListItemFileIcon } from '@components/shared/uploadFiles/uploadList/uploadListItem/uploadListItemFileIcon/uploadListItemFileIcon.component';
 import { UploadListItemTitle } from '@components/shared/uploadFiles/uploadList/uploadListItem/uploadListItemTitle/uploadListItemTitle.component';
